Solve settings
